pyproject.toml
[project]
name = "nexustrader-ai"
version = "1.0.0"
description = "AI-Powered Autonomous Trading Bot with Telegram Interface"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "NexusTrader Team"}
]
keywords = ["trading", "bot", "ai", "telegram", "binance", "crypto"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "ccxt>=4.0.0",
    "python-telegram-bot>=20.0",
    "openai>=1.0.0",
    "streamlit>=1.25.0",
    "plotly>=5.0.0",
    "supabase>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
]
ml = [
    "tensorflow>=2.13.0",
    "stable-baselines3>=2.0.0",
    "hmmlearn>=0.3.0",
]

[project.urls]
Homepage = "https://github.com/nexustrader/nexustrader-ai"
Documentation = "https://github.com/nexustrader/nexustrader-ai#readme"
Repository = "https://github.com/nexustrader/nexustrader-ai.git"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | \.mypy_cache
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["modules", "dashboard"]

[tool.flake8]
max-line-length = 100
extend-ignore = ["E501", "W503", "E203"]
exclude = [".git", "__pycache__", ".venv", "*.egg-info"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [".venv", "__pycache__"]

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["."]
omit = ["test_*.py", ".venv/*", "__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
